Clean up resources and scripts
